third_party/WebKit/LayoutTests/virtual/paint-timing/external/wpt/paint-timing/idlharness.window-expected.txt

This is a testharness.js-based test.
FAIL idl_test setup promise_test: Unhandled rejection with value: "Timed out waiting for paint event"
PASS PerformancePaintTiming interface: existence and properties of interface object
PASS PerformancePaintTiming interface object length
PASS PerformancePaintTiming interface object name
PASS PerformancePaintTiming interface: existence and properties of interface prototype object
PASS PerformancePaintTiming interface: existence and properties of interface prototype object's "constructor" property
PASS PerformancePaintTiming interface: existence and properties of interface prototype object's @@unscopables property
FAIL PerformancePaintTiming must be primary interface of paintTiming assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: paintTiming is not defined"
FAIL Stringification of paintTiming assert_equals: Unexpected exception when evaluating object expected null but got object "ReferenceError: paintTiming is not defined"
Harness: the test ran to completion.
